Updated task.

+++ Tasks/01 - Variables and Arithmetic/117-typecasting/117-typecasting.cpp
@@ -18,6 +18,6 @@ int main()
 	//This also needs fixing
 	int p = 123;
 	int q = 2;
-	double y = p / q;
+	double y = double(p) / double(q);
 	printf("y = %f\n", y);
 }
